+++ lib/gas/kinetics/energy-exchange-rate.cxx
@@ -25,7 +25,7 @@ Energy_exchange_rate(lua_State *L)
 	input_error(ost);
     }
     int nmechs = lua_objlen(L, -1);
-    for ( size_t i = 1; i <= nmechs; ++i ) {
+    for ( int i = 1; i <= nmechs; ++i ) {
 	lua_rawgeti(L, -1, i);
 	ee_mech_.push_back(create_energy_exhange_mechanism(L, imode));
 	lua_pop(L, 1);
